my_project/requirements_backend.txt
fastapi[standard]
Pillow
uvicorn
numpy
onnxruntime
pydantic
google-cloud-storage
